% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMatdec.plot2d.R
\name{myMatdec.plot2d}
\alias{myMatdec.plot2d}
\title{myMatdec.plot2d}
\usage{
myMatdec.plot2d(X)
}
\arguments{
\item{X}{a 3x2 matrix}
}
\value{
a plot of the data matrix and the mean
}
\description{
An S3 method for `myMatdec` which plots the data matrix and mean in 2D
}
\examples{
\dontrun{myMatdec.plot2d(X=X)}
}
