% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMatdec.R
\name{myMatdec}
\alias{myMatdec}
\title{myMatdec}
\usage{
myMatdec(X)
}
\arguments{
\item{X}{a 3x2 matrix}
}
\value{
a named list with various items of not-so-nefarious natures
}
\description{
A function that breaks down a 3x2 matrix into important parts for OU course MATH5793
}
\examples{
\dontrun{myMatdec(X=X)}
}
